reorganize file structure for components/containers
