+++ package.json
@@ -13,11 +13,13 @@
     "react": "^16.9.0",
     "react-browser-router": "^2.1.2",
     "react-dom": "^16.9.0",
-    "react-redux": "^7.1.0",
+    "react-redux": "^7.1.1",
+    "react-router": "^5.0.1",
     "react-router-dom": "^5.0.1",
     "react-scripts": "3.1.1",
     "redux": "^4.0.4",
-    "redux-devtools-extension": "^2.13.8"
+    "redux-devtools-extension": "^2.13.8",
+    "typescript": "^3.6.2"
   },
   "scripts": {
     "start": "react-scripts start",
