--- src/ProjectContainer/ProjectContainer.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const ProjectContainer = () => {
-  return (
-    <section>
-      
-    </section>
-  )
-}
-
-export default ProjectContainer;
