--- src/SingleProject/SingleProject.test.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-// import App from './App';
-import { shallow } from 'enzyme'
-
-describe('Project Container', () => {
-  it('should do something', () => {
-
-  })
-})
