+++ src/components/Color/Color.js
