+++ src/components/Color/Color.test.js
