+++ src/containers/NewPaletteForm/NewPaletteForm.test.js
