+++ src/containers/Palette/Palette.test.js
